Recommitting Version with Ints

+++ main.py
@@ -1,8 +1,8 @@
 
 def main():
-    num1 = float(input('Enter the first number: '))
-    num2 = float(input('Enter the second number: '))
-    num3 = float(input('Enter the third number: '))
+    num1 = int(input('Enter the first number: '))
+    num2 = int(input('Enter the second number: '))
+    num3 = int(input('Enter the third number: '))
   
     minval = int()
     maxval = int()
